Fix: Make lowfat check the full range of the access instead of just one-past the highest accessed address

+++ lowfat/src/check.h
@@ -5,7 +5,7 @@
 
 uint64_t __lowfat_ptr_index(void *ptr)
     __attribute__((__weak__, __always_inline__));
-uint64_t __lowfat_ptr_base(void *ptr, uint64_t index)
+uintptr_t __lowfat_ptr_base(void *ptr, uint64_t index)
     __attribute__((__weak__, __always_inline__));
 uint64_t __lowfat_ptr_size(uint64_t index)
     __attribute__((__weak__, __always_inline__));
